[models] Caused Subnet._str_ to return name #74
The AbstractSubnet returns a name and subnet if a name if supplied and
just a subnet if no name is supplied
To test this, a test was created. the test fails if the subnet str
doesn't return the name and subnet when a name and subnet is given else
it passes.

Fixes #74